+++ pykechain/__init__.py
@@ -8,9 +8,9 @@ from .helpers import get_project
 __all__ = ("Client", "get_project", "version")
 
 if sys.version_info.major == 2 or (
-    sys.version_info.major == 3 and sys.version_info.minor < 6
+    sys.version_info.major == 3 and sys.version_info.minor < 7
 ):
     raise RuntimeError(
-        "Python version >= `3.6` is required for this version of `pykechain` to operate. "
-        "Please use `pykechain` version `2.7.x` for usage in combination with Python `2.7`"
+        "Python version >= `3.7` is required for this version of `pykechain` to operate. "
+        "Please use `pykechain` version `3.x.x` for usage in combination with Python `3.6`"
     )